% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TaxSEA-package.R
\docType{package}
\name{TaxSEA-package}
\alias{TaxSEA-package}
\title{TaxSEA: Taxon Set Enrichment Analysis}
\description{
TaxSEA makes metagenomic data easier to interpret. Rather than reading species one at a time, it tests whether a-priori defined sets of taxa shift together, drawing on public databases (BugSigDB, MiMeDB, GutMGene, mBodyMap, BacDive and GMRepoV2) and on sets collated from the literature. Two approaches are provided. TaxSEA takes taxon names or NCBI IDs with a rank per taxon, such as a fold change or correlation coefficient, and tests each set with a Kolmogorov-Smirnov test. ssTaxSEA scores every set in a single sample as the mean centered log-ratio of its members, giving a per-sample value to use in any downstream model. Both work with any taxonomic profiling technology, including 16S rRNA gene sequencing, shotgun metagenomics and metatranscriptomics.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/feargalr/taxsea}
  \item \url{https://feargalr.github.io/TaxSEA/}
  \item Report bugs at \url{https://github.com/feargalr/taxsea/issues}
}

}
\author{
\strong{Maintainer}: Feargal Ryan \email{feargalr@gmail.com} (\href{https://orcid.org/0000-0002-1565-4598}{ORCID}) (funding: Supported by NHMRC Investigator Grant) [funder]

Authors:
\itemize{
  \item Feargal Ryan \email{feargalr@gmail.com} (\href{https://orcid.org/0000-0002-1565-4598}{ORCID}) (funding: Supported by NHMRC Investigator Grant) [funder]
}

}
\keyword{internal}
