% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TaxSEA.R
\name{TaxSEA_barcode}
\alias{TaxSEA_barcode}
\title{Create a barcode plot of TaxSEA results}
\usage{
TaxSEA_barcode(
  taxsea_results,
  taxon_ranks,
  axis_limits = c(-7, 7),
  n_to_plot = 10,
  boxplot = FALSE
)
}
\arguments{
\item{taxsea_results}{A data frame containing the results of the 
TaxSEA function}

\item{taxon_ranks}{A named numeric vector with taxon names as names 
and their corresponding ranks as values}

\item{axis_limits}{A numeric vector of length 2 specifying the x-axis
limits (default: c(-7, 7))}

\item{n_to_plot}{Integer value specifying the number of top taxon
sets to display (default: 10)}

\item{boxplot}{Boolean value specifying whether or not to include a 
boxplot (default: FALSE)}
}
\value{
A ggplot2 barcode plot of the TaxSEA results
}
\description{
This function takes a TaxSEA result data frame, taxon ranks, and taxon 
sets to create a barcode plot
using ggplot2. The barcode plot shows the taxon sets with the top 
significant P values.
}
