% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TaxSEA.R
\name{TaxSEA_barplot}
\alias{TaxSEA_barplot}
\title{Create a bar plot of TaxSEA results}
\usage{
TaxSEA_barplot(taxsea_results, threshold = 0.2, custom_colors = NULL)
}
\arguments{
\item{taxsea_results}{A data frame containing the results of the TaxSEA 
function}

\item{threshold}{Numeric value representing the FDR threshold for displaying 
taxon sets (default: 0.2)}

\item{custom_colors}{A character vector of length 2 with colors for 
upregulated and downregulated taxon sets, respectively (default: NULL)}
}
\value{
A ggplot2 bar plot of the TaxSEA results
}
\description{
This function takes a TaxSEA result data frame and creates a bar plot of 
the results
using ggplot2. It highlights upregulated and downregulated taxon sets 
based on their #' normalized enrichment scores (NES) and false discovery 
rate (FDR).
}
