% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_taxon_sets.R
\name{get_taxon_sets}
\alias{get_taxon_sets}
\title{Retrieve Taxon Sets from TaxSEA Library}
\usage{
get_taxon_sets(taxon_to_fetch = taxon)
}
\arguments{
\item{taxon_to_fetch}{The taxon to search for in the TaxSEA database.}
}
\value{
A character vector containing the names of taxonomic sets where the specified taxon is present.
}
\description{
Retrieve from the TaxSEA database which taxon sets (metabolite producers and disease signatures) contain a taxon of interest.
}
\examples{
# Retrieve sets for Bifidobacterium longum
get_taxon_sets(taxon="Bifidobacterium_longum")

}
