% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TaxSEA.R
\name{NCBI_ids}
\alias{NCBI_ids}
\title{NCBI IDs and Taxonomic Aliases}
\format{
A named vector where names are taxonomic aliases and values are NCBI taxonomic identifiers.
}
\usage{
data("NCBI_ids")
}
\description{
A named vector that provides a mapping between NCBI taxonomic identifiers and their corresponding taxonomic aliases.
}
