% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NCBI_ids}
\alias{NCBI_ids}
\title{NCBI IDs Dataset}
\format{
A named vector where:
\describe{
  \item{names}{NCBI IDs}
  \item{values}{Species or genus names}
}
}
\source{
NCBI
}
\usage{
NCBI_ids
}
\description{
A dataset for mapping NCBI IDs to species/genus names. This named vector
allows for lookup of NCBI IDs associated with species or genus names.
}
\examples{
data(NCBI_ids)
# Can look up either with or without spaces
NCBI_ids["Bifidobacterium_breve"]
NCBI_ids["Bifidobacterium breve"]
}
\keyword{datasets}
