% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TaxSEA-package.R
\docType{package}
\name{TaxSEA-package}
\alias{TaxSEA-package}
\title{TaxSEA: Taxon Set Enrichment Analysis}
\description{
TaxSEA is an R package for Taxon Set Enrichment Analysis, which utilises a Kolmogorov-Smirnov test analyses to investigate differential abundance analysis output for whether there are alternations in a-priori defined sets of taxa from public databases (BugSigDB, MiMeDB, GutMGene, mBodyMap, BacDive and GMRepoV2) and collated from the literature. TaxSEA takes as input a list of taxonomic identifiers (e.g. species names, NCBI IDs etc.) and a rank (E.g. fold change, correlation coefficient). TaxSEA be applied to any microbiota taxonomic profiling technology (array-based, 16S rRNA gene sequencing, shotgun metagenomics & metatranscriptomics etc.) and enables researchers to rapidly contextualize their findings within the broader literature to accelerate interpretation of results.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/feargalr/taxsea}
  \item \url{https://feargalr.github.io/TaxSEA/}
  \item Report bugs at \url{https://github.com/feargalr/taxsea/issues}
}

}
\author{
\strong{Maintainer}: Feargal Ryan \email{feargalr@gmail.com} (\href{https://orcid.org/0000-0002-1565-4598}{ORCID}) (funding: Supported by NHMRC Investigator Grant) [funder]

Authors:
\itemize{
  \item Feargal Ryan \email{feargalr@gmail.com} (\href{https://orcid.org/0000-0002-1565-4598}{ORCID}) (funding: Supported by NHMRC Investigator Grant) [funder]
}

}
\keyword{internal}
