% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TaxSEA.R
\name{TaxSEA}
\alias{TaxSEA}
\title{TaxSEA: Taxon Set Enrichment Analysis}
\usage{
TaxSEA(taxon_ranks, database = "All")
}
\arguments{
\item{taxon_ranks}{A named vector of log2 fold changes between control and test group}

\item{database}{A character specifying the database to use for enrichment analysis.
Options are "All", "GutMGene", "MiMeDB", and "GMRepoV2". Default is "All".}
}
\value{
A data frame with taxon set enrichment results
}
\description{
TaxSEA is designed to enable rapid annotation of changes observed in a microbiome association study
by testing for enrichment for producers of particular metabolites, or associations with marker taxa
for particular diseases. It focuses specifically on human gut microbiome associations and uses
a Kolmogorov-Smirnov test to test if a particular set of taxa is changed relative to a control group.
The input taxon_ranks are log2 fold changes between control and test group (e.g. healthy and IBD).
}
\examples{
# Example data
taxon_ranks <- runif(10, -3, 3)
names(taxon_ranks) <- paste("Taxon", 1:10)
taxon_sets <- list(
  set1 = c("Taxon 1", "Taxon 2", "Taxon 3"),
  set2 = c("Taxon 4", "Taxon 5", "Taxon 6"),
  set3 = c("Taxon 7", "Taxon 8", "Taxon 9")
)
# Run TaxSEA
result_df <- TaxSEA(taxon_ranks)
}
\seealso{
\itemize{
  \item \url{https://doi.org/10.1093/nar/gkac868} for MiMeDB
  \item \url{https://doi.org/10.1093/nar/gkab1019} for GMrepo
  \item \url{https://doi.org/10.1093/nar/gkab786} for gutMGene
}
}
