% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TaxSEA.R
\name{TaxSEA_db}
\alias{TaxSEA_db}
\title{TaxSEA Database}
\format{
A list where each element is a vector. Element names are taxon set names, and vector elements are the members of those sets.
}
\source{
Your source here (if applicable)
}
\usage{
data("TaxSEA_db")
}
\description{
A named list of vectors, where each name corresponds to a taxon set, and the associated vector contains the members of that set.
}
