% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TaxSEA.R
\name{TaxSEA_test_data}
\alias{TaxSEA_test_data}
\title{TaxSEA Example Input Data}
\format{
A numeric vector of IDs.
}
\source{
Your source here (if applicable)
}
\usage{
data("TaxSea_test_data")
}
\description{
An example input vector for TaxSEA, consisting of numeric values representing IDs.
}
