% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{TaxSEA_test_data}
\alias{TaxSEA_test_data}
\title{TaxSEA Test Data}
\format{
A data frame with two columns:
\describe{
  \item{rank}{Character vector representing taxon ranks}
  \item{id}{Character vector representing taxon IDs}
}
}
\source{
See READ ME.
}
\description{
A dataset containing taxon ranks and taxon IDs.
}
\examples{

data(TaxSEA_test_data)
test_results <- TaxSEA(TaxSEA_test_data)
}
\value{
A vector containing the ranks and species IDs
}
